Wait, what is my time complexity?
O(m^n)... where m is the number of words from which it draws (~250 000), and n is the number of slots, which goes up to 100 or so.

Okay, that's not entirely true. m is really only the number of words of a given length, which only goes up to more like 10 000. And n is actually the number of backtracks, which can easily be kept below 2 or 3 if you're lucky. So it's a brute force algorithm, yes, but an optimised one. It actually works quite nicely if you just run it for 10 seconds, cancel it if it doesn't succeed, and try again.

I've made some changes to speed it up as much as I can without actually improving the backtracking. And made the layout easier for it. And it works! I have a puzzle! So I'm going to stop working on this now. Muahaha!